Make some changes on top of @mlehner's changes suggested by @stof in https://github.com/Seldaek/monolog/pull/61
I modified composer.json to point to my fork of gelf-php
(https://github.com/msabramo/gelf-php) for the time being, as my fork
has a composer.json, namespaces, and the IMessagePublisher interface.
I've also submitted pull requests with my changes upstream to
https://github.com/Graylog2/gelf-php. If those changes are accepted,
then we can use the official repo or if they register the package on
Packagist (which I suggested in one of my requests), then we can just
eliminate the repository altogether.