Adding a first version of PushoverHandler
The PushoverHandler sends messages to api.pushover.net. The class depends on the SocketHandler
I modified the SocketHandler to make it possible to change the port for subclasses (Pushover needs the explicit port 80 to work)
The tests only see if the content sent to the api is correct. The rest should be already covered in the tests for the Sockets.