Use `text` and `title` fields for SlackHandler
According to slack documentation (https://api.slack.com/docs/attachments),  an attachment may have a title and a text. The log message and title should use these fields. One of the benefits of using these two properties instead of adding a field is that it's shortened (with a "show more" button) and expandable by slack. It adds a bit of hierarchy in the attachement. Message is the content. Extra and context are additional fields.