Provide means to manually activate a FingersCrossedHandler
This is useful if you have a FingersCrossedHandler set up and some where
in your application are interested in all output regardless of whether
the threshold is actually reached. For example to temporarily debug some
problem. One can replace the handler in that spot, but this is
signifcantly easier to use.