Makes tests more explicit
This commit adds tests which showcase a use case for
a parent logger which cannot otherwise be achieved. The specific
case is the following.
ParentLogger:
    Handlers: [ParentNonBubblingWarningHandler, ParentBubblingNoticeHandler]
ChildLogger:
    Handlers: [ChildNonBubblingCriticalHandler, ChildBubblingInfoHandler]

In the above, messages issues with the following levels go to the
following handlers.

Degbug:

Info:
    ChildBubblingInfoHandler
Notice:
    ChildBubblingInfoHandler, ParentBubblingNoticeHandler
Warning:
    ChildBubblingInfoHandler, ParentNonBubblingWarningHandler
Error:
    ChildBubblingInfoHandler, ParentNonBubblingWarningHandler
Critical:
    ChildNonBubblingCriticalHandler, ParentNonBubblingWarningHandler
Alert:
    ChildNonBubblingCriticalHandler, ParentNonBubblingWarningHandler
Emergency:
    ChildNonBubblingCriticalHandler, ParentNonBubblingWarningHandler

This may be the case when a certain service instantiates a logger and
passes it into the constructor of another class. In this case, the
service's logger is the parent logger and the instantiated class may
instantiate its own logger (the child) and use the passed in logger as
the parent. This way all messages logged through any class used by the
service will be logged per the parent logger's specs. However,
individual classes can have some additional control over their own
logging.