fix endless recursion of normalizer formatter
If the NormalizerFormatter encounters an instance of Traversable it
calls itself on all children of that instance. If the Traversable
returns itself (once, sometimes or always -- doesn't matter) the
formatter goes into endless recursion.

Fix by checking the currently being normalized data to not be
identity-equal to its parent container.