Add Hostname Processor
I have found myself reimplementing this processor a number of times for
different clients. This processor is useful when you're running a number
of instances of an application across multiple hosts. This allows you to
narrow incidents to a specific host.

This will add the hostname to extra, under the "hostname" key. On
failure to retrieve this value will be false.