Prevent to call newrelic functions when app and transaction names are empty
The handler shouldn't set the app and transaction names when they are not been passed in the constructor or provided through the context. If the value returned by `getAppName()` or `getTransactionName()` is null or an empty string, the correspondent New Relic API shouldn't be called because the user might have called them in another part of the code. And anyway it doesn't make any sense call `newrelic_set_appname()` and `newrelic_name_transaction()` passing an empty string or null.