Update LineFormatter.php replaceNewlines {
Hello,

When enabling allowInlineLineBreaks and there is no { in the string, it returns a FALSE instead of a 0 so the function will not get triggered. I have filled in { in the first character of my string to fix this but i think it must be:

```
if (false === strpos($str, '{')) {
```
Instead of 
```
if (0 === strpos($str, '{')) {
```

or am i mistaken?