We do not need to close the socket manually
PHP docs say this about __destruct(): "The destructor method will be called as
soon as there are no other references to a particular object, or in any order
during the shutdown sequence.". Monolog will attempt to close the socket via
AbstractHandler::__destruct() but if this occurs during the PHP "shutdown
sequence", it seems a send() may actually come after this which triggers an
error: "The UdpSocket to syslog.cbxone:514 has been closed and can not be
written to anymore".

Using strace I have verified that PHP closes the socket itself during shutdown,
even with FPM (which reuses PHP processes), so the __destruct behavior is
entirely unnecessary