Improve exception message
This exception is commonly thrown in Laravel applications when using Docker.
But it is not clear why the system could not open a file that has read permission
for all users. So, when I came to the code of Monolog I saw that you try to open it
in append mode, so I think that just adding this information could help a lot other
users.