Reduce calls to date_default_timezone_get()
Not sure why, but this call seems to be time consuming enough to appear
in performance profiles (maybe reading from disk is involved or
something).
This should probably not be resolved in dependency injection containers
since one might one to build the container on one machine, and then
deploy on many machines in different timezones.