Check if errorMessage  contain error "File exists"
Fixes #1678, closes #1685

When we try to create directory we got error and find out that error is to the fact that directory already was created for us.
If that the case we should not throw exception as it's fine now...
If file was deleted after that it's not problem of this funtion.